% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ela1}
\alias{ela1}
\title{Mixed-Format English Language Arts (ELA) Assessment Data (Form 1)}
\format{
A data frame with 52417 rows and 47 columns:
\itemize{
\item \code{ITEM1} - \code{ITEM30}: Binary responses (0 = Incorrect, 1 = Correct).
\item \code{ITEM31} - \code{ITEM45}: Polytomous responses (scored 0-5).
\item \code{THETA}: Latent ability estimates.
\item \code{COVARIATE}: Person-level background variable.
}
}
\source{
Tang, C., Xiong, J., & Engelhard, G. (2025). Identification of writing
strategies in educational assessments with an unsupervised learning
measurement framework. \emph{Education Sciences}, 15(7), 912.
\doi{10.3390/educsci15070912}
}
\usage{
ela1
}
\description{
A dataset containing binary and polytomous responses for demonstration.
}
\examples{
data(ela1)
head(ela1)
}
\keyword{datasets}
