% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ela2}
\alias{ela2}
\title{Mixed-Format English Language Arts (ELA) Assessment Data (Form 2)}
\format{
A data frame with columns representing item responses.
\itemize{
\item \code{ITEM1} - \code{ITEM7}: Binary responses (0 = Incorrect, 1 = Correct).
\item \code{ITEM8}: Polytomous response (scored 0-2).
\item \code{ITEM9}: Polytomous response (scored 0-5).
\item \code{ITEM10}: Polytomous response (scored 0-5).
}
}
\source{
Tang, C., Xiong, J., & Engelhard, G. (2025). Identification of writing
strategies in educational assessments with an unsupervised learning
measurement framework. \emph{Education Sciences}, 15(7), 912.
\doi{10.3390/educsci15070912}
}
\usage{
ela2
}
\description{
A smaller dataset containing item responses.
}
\examples{
data(ela2)
head(ela2)
}
\keyword{datasets}
