% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ela3}
\alias{ela3}
\title{Large-Scale Mixed-Format English Language Arts (ELA) Assessment Data (Form 3)}
\format{
A data frame with 2,434,185 observations and 5 variables:
\itemize{
\item{STUDENTID}: Unique student identifier.
\item{FORMID}: Unique test form identifier.
\item{ITEMID}: Unique item identifier, where _T1 indicates trait 1 and _T2 indicates trait 2.
\item{SEQ}: Position of the item in a form.
\item{SCORE}: Dichotomously and polytomously scored item responses (0,1,2...).
}
}
\source{
Tang, C., Xiong, J., & Engelhard, G. (2025). Identification of writing
strategies in educational assessments with an unsupervised learning
measurement framework. \emph{Education Sciences}, 15(7), 912.
\doi{10.3390/educsci15070912}
}
\usage{
ela3
}
\description{
A long format dataset containing binary and polytomous responses for demonstration.
}
\examples{
data(ela3)
head(ela3)
}
\keyword{datasets}
