% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ela3_testmap}
\alias{ela3_testmap}
\title{Large-Scale Mixed-Format English Language Arts (ELA) Assessment Data Testmap (Form 3 Testmap)}
\format{
A data frame with 328 rows and 5 variables:
\itemize{
\item{FORMID}: Unique test form identifier.
\item{SEQ}: Position of the item in a form.
\item{ITEMID}: Unique item identifier, where _T1 indicates trait 1 and _T2 indicates trait 2.
\item{TYPE}: Type of an item, where OP indicates operational items and FT indicates field-test items.
\item{MAX_SCORE}: Max score of that item.
}
}
\source{
Tang, C., Xiong, J., & Engelhard, G. (2025). Identification of writing
strategies in educational assessments with an unsupervised learning
measurement framework. \emph{Education Sciences}, 15(7), 912.
\doi{10.3390/educsci15070912}
}
\usage{
ela3_testmap
}
\description{
Test map for the long format ela3 response data
}
\examples{
data(ela3_testmap)
head(ela3_testmap)
}
\keyword{datasets}
