% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{expected.test_mod}
\alias{expected.test_mod}
\title{modified expected.test()}
\usage{
expected.test_mod(x, Theta, which.items, mydata)
}
\arguments{
\item{mydata}{dataframe from which category values of individual items are extracted}
}
\description{
used within `tirt()` to computeexpected score
modified from mirt::expected.test() to account for different category values
}
