% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tirt.R
\name{tirt}
\alias{tirt}
\title{Threshold based on IRT method (tirt)}
\usage{
tirt(mydata, B = 0, rform = c("prop", "sum"))
}
\arguments{
\item{mydata}{a dataframe containing all items of a multi-item questionnaire (or Frailty Index) and a binary anchor item}

\item{B}{number of bootstrap resamples to do to get confidence intervals for IRT-based thresholds}

\item{rform}{computes expected summed score (`sum`) or express summed score as a proportion of number of items (`prop` default)}
}
\value{
a list of results comprising the expected scores (expressed as a proportion or sum) and bootstrapped CIs if B > 0
}
\description{
Threshold based on IRT method (tirt)
}
\examples{
ensure mirt package is installed
Ensure dataset has no missing values
identify all vars with non-integer values
ensure binary anchor item
convert non-integers to integers
B > 0, bootstrap IRT-based threshold
convert non-integers to integers
}
