% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{yhcheck}
\alias{yhcheck}
\title{item response option checks}
\usage{
yhcheck(mydata, n_unique = 1)
}
\arguments{
\item{n_unique}{return vector of vars with <=`nunique` values}
}
\description{
Used within `tirt()`
checks whether items have 2 or more response options
important to perform checks on low-prevalence items during the bootstrapping process
}
